requirements.txt
numpy==1.22.2
opencv_python==4.5.5.62
Pillow==9.0.1
pygrabber==0.1
scipy==1.8.0
TkinterDnD2==0.3.0
tkinterdnd2_pmgagne==0.3.0
torch==1.10.2+cu102
torchvision==0.11.3
